services/system/api/requirements.txt
aioredis==2.0.1
anyio==3.6.2
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.27.0
black==22.10.0
certifi==2022.9.24
click==8.1.3
fastapi==0.87.0
flake8==5.0.4
h11==0.14.0
httpcore==0.16.1
httptools==0.5.0
httpx==0.23.1
idna==3.4
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.10.2
platformdirs==2.5.4
psutil==5.9.4
py-cpuinfo==9.0.0
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
PyJWT==2.6.0
python-dotenv==0.21.0
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.21.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
